+++ packages/fast-components/src/index.ts
@@ -1,4 +1,6 @@
+export * from "./anchor";
 export * from "./badge";
+export * from "./button";
 export * from "./card";
 export * from "./checkbox";
 export * from "./design-system-provider";