+++ packages/fast-figma-plugin-msft/src/main.ts
@@ -1,5 +1,11 @@
 import { PluginNode, PluginNodeData } from "./core/node";
-import { fillRecipes, RecipeStore, strokeRecipes, textFillRecipes } from "./core/recipes";
+import {
+    cornerRadiusRecipe,
+    fillRecipes,
+    RecipeStore,
+    strokeRecipes,
+    textFillRecipes,
+} from "./core/recipes";
 import { FigmaController } from "./figma/controller";
 import { RecipeDefinition, RecipeTypes } from "./core/recipe-registry";
 import { canHaveChildren, FigmaPluginNode, isInstanceNode } from "./figma/node";
@@ -7,21 +13,13 @@ import { MessageTypes, UIMessage } from "./core/messaging";
 
 const controller = new FigmaController();
 
-function friendlyName(str: string): string {
-    const result = str
-        .split(/([A-Z])/)
-        .reduce((prev: string, current: string, index: number) => {
-            return index % 2 === 1 ? prev + " " + current : prev + current;
-        });
-
-    return result.charAt(0).toUpperCase().concat(result.slice(1));
-}
-
-function register(type: RecipeTypes, recipes: RecipeStore): void {
+function registerColorRecipe(type: RecipeTypes, recipes: RecipeStore): void {
     Object.keys(recipes).forEach((key: string) => {
+        const recipe = recipes[key];
+
         const definition: RecipeDefinition = {
             id: key,
-            name: friendlyName(key),
+            name: recipe.name,
             type,
             evaluate: (node: PluginNode): string => {
                 const parent = node.parent();
@@ -29,7 +27,7 @@ function register(type: RecipeTypes, recipes: RecipeStore): void {
                     ? parent.getEffectiveBackgroundColor()
                     : node.getEffectiveBackgroundColor();
 
-                return recipes[key]({
+                return recipe.resolver({
                     ...node.designSystem,
                     backgroundColor: backgroundColor.toStringHexRGB(),
                 } as any);
@@ -65,15 +63,30 @@ function syncInstanceWithMaster(target: InstanceNode): void {
 /**
  * Register recipe types
  */
-register(RecipeTypes.backgroundFills, fillRecipes);
-register(RecipeTypes.foregroundFills, textFillRecipes);
-register(RecipeTypes.strokeFills, strokeRecipes);
+registerColorRecipe(RecipeTypes.backgroundFills, fillRecipes);
+registerColorRecipe(RecipeTypes.foregroundFills, textFillRecipes);
+registerColorRecipe(RecipeTypes.strokeFills, strokeRecipes);
 
+Object.keys(cornerRadiusRecipe).forEach(key => {
+    const recipe = cornerRadiusRecipe[key];
+
+    const definition: RecipeDefinition = {
+        id: key,
+        name: recipe.name,
+        type: RecipeTypes.cornerRadius,
+        evaluate: (): number => {
+            return recipe.resolver({} as any);
+        },
+    };
+
+    controller.recipeRegistry.register(definition);
+});
 /**
  * Show UI on plugin launch
  */
 figma.showUI(__html__, {
     height: 600,
+    width: 356,
 });
 
 /**