+++ packages/fast-figma-plugin-msft/src/ui.tsx
@@ -1,8 +1,13 @@
 import React from "react";
 import ReactDOM from "react-dom";
+import { FASTDesignSystemProvider } from "@microsoft/fast-components";
 import { PluginUI, PluginUIProps } from "./core/ui";
 import { UIMessage } from "./core/messaging";
 
+/* eslint-disable */
+const styles = require("./global.css");
+FASTDesignSystemProvider;
+/* eslint-enable */
 /**
  * Dispatches a UI message to the host
  * @param message The message to dispatch
@@ -11,19 +16,11 @@ function dispatchMessage(message: UIMessage): void {
     parent.postMessage({ pluginMessage: message }, "*");
 }
 
+const root = document.querySelector("fast-design-system-provider");
+
 function render(props?: PluginUIProps): void {
     ReactDOM.render(<PluginUI {...props} dispatch={dispatchMessage} />, root);
 }
-
-// Import with require so the dependency doesn't get tree-shaken
-const styles = require("./global.css");
-
-/**
- * Create root element
- */
-const root: HTMLDivElement = document.createElement("div");
-root.setAttribute("id", "root");
-document.body.appendChild(root);
 
 /**
  * Wire iframe's onmessage to render function