+++ packages/fast-figma-plugin-msft/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-figma-plugin-msft@0.4.7...@microsoft/fast-figma-plugin-msft@0.5.0) (2020-04-13)
+
+
+### Features
+
+* update figma plugin UI ([#2912](https://github.com/Microsoft/fast-dna/issues/2912)) ([aa4f57d](https://github.com/Microsoft/fast-dna/commit/aa4f57de6660924dcb910352f21e78d7fe2bba34))
+
+
+
+
+
 ## [0.4.7](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-figma-plugin-msft@0.4.6...@microsoft/fast-figma-plugin-msft@0.4.7) (2020-04-10)
 
 **Note:** Version bump only for package @microsoft/fast-figma-plugin-msft